Added a confirmation alert

+++ src/components/REST/REST Account/RestAccount.tsx
@@ -49,7 +49,7 @@ export const RestAccount = ():[ (userid:number)=> void,(userid:number)=>void,(ac
     }
 
     function editAccount(account:Account){
-        axios.put("http://localhost:8080/product/getByProduct?productid=" + account.userid,{
+        axios.put("http://localhost:8080/user/putUser?userid=" + account.userid,{
             username: account.username,
             password: account.password,
             account_Type: account.account_type,
@@ -71,27 +71,31 @@ export const RestAccount = ():[ (userid:number)=> void,(userid:number)=>void,(ac
         })
     }
 
-    function getAccountbyId (userid:number|undefined){
-        axios.get("http:localhost:8080/user/getByUser?userid=" + userid,{
-    }).then((response) => {
-        setAccount(response.data);
-        console.log(response.data);
-    })
-    .catch((error) => {
-        setError(error.message);
-    })
+    function getAccountbyId (userid:number|undefined) {
+        axios.get("http://localhost:8080/user/getByUser?userid=" + userid,{
+        }).then((response) => {
+            setAccount(response.data);
+            console.log(response.data);
+        })
+        .catch((error) => {
+            setError(error.message);
+        })
     }
 
-function deleteByID (userid:number|undefined){
-    axios.delete("http://localhost:8080/user/deleteAccount/" + userid,{
-    }).then((response) => {
-    setAccount(response.data);
-    console.log(response.data);  
-    window.location.reload()  
-    })
-.catch((error) => {
-    setError(error.message);
-    })
-}
+    function deleteByID (userid:number|undefined){
+        const confirm = window.confirm(`Are you sure you want to delete the account for this account?`);
+        if (confirm) {
+            axios.delete("http://localhost:8080/user/deleteAccount/" + userid,{
+            }).then((response) => {
+                setAccount(response.data);
+                console.log(response.data);  
+                window.location.reload()  
+            })
+            .catch((error) => {
+                setError(error.message);
+            })
+        }
+        
+    }
     return[deleteByID,getAccountbyId,editAccount,addAccount,account,error]
 }
